+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/questions/BgpProcessPropertySpecifier.java
@@ -17,7 +17,6 @@ import javax.annotation.Nullable;
 import org.batfish.datamodel.BgpActivePeerConfig;
 import org.batfish.datamodel.BgpPassivePeerConfig;
 import org.batfish.datamodel.BgpProcess;
-import org.batfish.datamodel.answers.AutocompleteSuggestion;
 import org.batfish.datamodel.answers.Schema;
 
 /**
@@ -86,14 +85,6 @@ public class BgpProcessPropertySpecifier extends PropertySpecifier {
     _expression =
         properties.stream().map(String::trim).map(Pattern::quote).collect(Collectors.joining("|"));
     _pattern = Pattern.compile(_expression, Pattern.CASE_INSENSITIVE);
-  }
-
-  /**
-   * Returns a list of suggestions based on the query, based on {@link
-   * PropertySpecifier#baseAutoComplete}.
-   */
-  public static List<AutocompleteSuggestion> autoComplete(String query) {
-    return PropertySpecifier.baseAutoComplete(query, JAVA_MAP.keySet());
   }
 
   @Override
