+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/questions/Variable.java
@@ -23,8 +23,6 @@ import org.batfish.datamodel.answers.AutocompleteSuggestion.CompletionType;
 public class Variable {
 
   public enum Type {
-    ADDRESS_BOOK("addressBook", true),
-    ADDRESS_GROUP("addressGroup", true),
     ANSWER_ELEMENT("answerElement", true),
     BGP_PEER_PROPERTY_SPEC("bgpPeerPropertySpec", true),
     BGP_PROCESS_PROPERTY_SPEC("bgpProcessPropertySpec", true),
